# Improve the performance of the getCellMeta calls