--- a/src/dataMap/index.js
+++ b/src/dataMap/index.js
@@ -1,5 +1,11 @@
 import DataMap from '../dataMap';
 import MetaManager from './metaManager';
 import metaSchemaFactory from './metaManager/metaSchema';
+import { DynamicCellMetaMod } from './metaManager/mods/dynamicCellMeta';
 
-export { DataMap, MetaManager, metaSchemaFactory };
+export {
+  DataMap,
+  MetaManager,
+  DynamicCellMetaMod,
+  metaSchemaFactory
+};