… .changelogs/6303-hooks.json
… .changelogs/6303.json
… src/dataMap/metaManager/__tests__/lazyFactoryMap.unit.js
… src/dataMap/metaManager/__tests__/utils.unit.js
… src/dataMap/metaManager/metaLayers/__tests__/cellMeta.unit.js
… src/dataMap/metaManager/metaLayers/__tests__/columnMeta.unit.js
… src/dataMap/metaManager/metaLayers/__tests__/tableMeta.unit.js
… src/dataMap/metaManager/metaLayers/globalMeta.js
… src/dataMap/metaManager/mods/__tests__/dynamicCellMeta.unit.js
… src/dataMap/metaManager/mods/dynamicCellMeta.js
… src/helpers/__tests__/array.unit.js
… src/helpers/__tests__/browser.unit.js
… src/helpers/__tests__/console.unit.js
… src/helpers/__tests__/data.unit.js
… src/helpers/__tests__/date.unit.js
… src/helpers/__tests__/feature.unit.js
… src/helpers/__tests__/function.unit.js
… src/helpers/__tests__/mixed.unit.js
… src/helpers/__tests__/number.unit.js
… src/helpers/__tests__/object.unit.js
… src/helpers/__tests__/string.unit.js
… src/helpers/__tests__/templateLiteralTag.unit.js
… src/helpers/dom/__tests__/element.unit.js
… src/helpers/dom/__tests__/event.unit.js
… src/helpers/function.js
… src/mixins/localHooks.js
… src/pluginHooks.js
… src/plugins/autoColumnSize/autoColumnSize.js
… src/plugins/autoRowSize/autoRowSize.js
… src/plugins/columnSorting/columnSorting.js