### Context
<!--- Why is this change required? What problem does it solve? -->
The PR aims to improve the performance of the `getCellMeta` calls by implementing an additional cache layer. The method is **hot** and it's called very often from different parts of the source code. The dynamic nature of the method adds the ability to extend the cell meta on the fly using the `cells` function and/or `beforeGetCellMeta` and `afterGetCellMeta` hooks. That causes performance bottlenecks.

From my research and observation (I tested the performance using [Perf Lab](https://github.com/handsontable/performance-lab)) adding a simple cache layer that is cleared once per slow render cycle ~or per index mapper change~ improves the table performance from ~1% to ~50% (or more). It all depends on how complicated the logic is implemented within the `cells` function or if the `"{after/before}GetCellMeta` hooks are used.

#### Breaking change

The cell meta cache layer is synchronized with the table render cycle. The cache is cleared once per the slow table render cycle. While developing the solution I hit the wall. It turned out that the hook `beforeRender` is triggered before the view rendering engine (Walkontable) renders the table not before the Handsontable renders the engine (see more explanation https://github.com/handsontable/handsontable/pull/8632). The hooks names now are more meaningful for developers and they are triggered in a more appropriate time in the table render life cycle. But that introduces the breaking change.
| Before (9.0.2) | After (PR) |
| ------------- | ------------- |
| `beforeRender`  | `beforeViewRender`  |
| `afterRender`  | `afterViewRender`  |
| - | `beforeRender`  |
| - | `afterRender`  |

The migration guide for the users would be a suggestion to rename all hooks within their apps from `beforeRender` or `afterRender` to new ones (`beforeViewRender` or `afterViewRender`).

#### The performance comparison of the `getCellMeta` method calls
1. `getCellMeta` calls for different Handsontable configurations:

![2021-08-20 o 13 17 58](https://user-images.githubusercontent.com/571316/130226550-e4e39653-07e1-4fd7-b5d4-e19f8b69b838.png)
The chart can be edited at https://live.amcharts.com/lNDc2/.

|  | 9.0.2 | 10.0.0 | Performance gain |
| ------------- | ------------- | ------------- | ------------- |
| "bare" HoT example, with enabled headers | 28.76 ms +-1% | 26.05 ms +-1% | ↑9.42 % |
| Using the `cells` option | 82.57 ms +-0% | 30.47 ms +-2% | ↑63.01 % |
| Using the `beforeGetCellMeta` hook | 40.16 ms +-1%  | 27.33 ms +-1%  | ↑31.94 % |
| Using the  `afterGetCellMeta` hook | 34.18 ms +-1% | 26.72 ms +-1% | ↑21.82 % |

2. `getCellMeta` calls for different dataset sizes:

![2021-08-20 o 13 17 37](https://user-images.githubusercontent.com/571316/130225662-114c5f20-f099-4b07-bdaa-7ffae18006f3.png)
(https://live.amcharts.com/djY2R/)

|  | 9.0.2 | 10.0.0 | Performance gain |
| ------------- | ------------- | ------------- | ------------- |
| 50x50 | 0.88 ms +-5% | 0.84 ms +-5% | ↑4.54 % |
| 200x100 | 6.06 ms +-3% | 5.28 ms +-3% | ↑12.87 % |
| 500x200 | 34.48 ms +-2%  | 26.88 ms +-3%  | ↑22.04 % |
| 8000x40 | 90.72 ms +-0% | 81.53 ms +-0% | ↑10.13 % |

The data was collected by sampling the same actions. The sampling size was set to 100. The tests ran on the same machine (MacBook Pro i7 2.3GHz 16GB).

#### How the changes impacts on the table usage (scrolling, editing, arrow keys navigation)
Using the simple "bare" example I didn't notice any changes in performance.
```javascript
new Handsontable(example, {
  data: Handsontable.helper.createSpreadsheetData(1000, 500),
  rowHeaders: true,
  colHeaders: true,
});
```
After adding some logic to the `cells` function that adds some class names to the cells and modifies the `readOnly` setting I am able to notice a more significant speedup. The result fluctuates between **1% to ~25%** (depends on the test case) faster script execution time comparing to the 9.0.2.

<details>
  <summary>`cells` example</summary>

  ```javascript
  new Handsontable(example, {
    data: Handsontable.helper.createSpreadsheetData(1000, 500),
    rowHeaders: true,
    colHeaders: true,
    cells(row, col) {
      const cellProperties = {};
      const visualRowIndex = this.instance.toVisualRow(row);
      const visualColIndex = this.instance.toVisualColumn(col);
      const data = this.instance.getDataAtCell(visualRowIndex, visualColIndex);

      if (visualRowIndex === 0 && visualColIndex === 0) {
        cellProperties.readOnly = true;
      }

      if (data.indexOf('5') >= 0) {
        cellProperties.className = 'red-bg';
      }

      if (row % 5 === 1) {
        cellProperties.className = 'yellow-bg';
      }

      if (col % 5 === 1) {
        cellProperties.className = 'blue-bg';
      }

      if (row % 12 === 1) {
        cellProperties.readOnly = true;
      }

      if (row % 7 === 1) {
        cellProperties.renderer = function(hot, td, row, col, prop, value) {
          td.textContent = `.x${value}x.`;
        };
      }

      return cellProperties;
    },
  });
  ```
</details>

Going further. In this example, I listening only to the `beforeGetCellMeta` hook to dynamically change the cell type to `date` depends on the visual row index. The improvements shape between **1% to ~8%**.

<details>
  <summary>`beforeGetCellMeta` hook example</summary>

  ```javascript
  new Handsontable(example, {
    data: Handsontable.helper.createSpreadsheetData(1000, 500),
    rowHeaders: true,
    colHeaders: true,
    beforeGetCellMeta(row, column, cellProperties) {
      if (row % 8 === 1) {
        cellProperties.type = 'date';
      }
    },
  });
  ```
</details>

I the example below I tried to put more heavy logic into the `cells`. Based on the previous example I put into the function more Handsontable API calls, I used the `moment` library to conditionally check the cell props and so on. The script execution time improves between **~15% to ~50%** (or more depends on how the logic within hooks or `cells` function is fast).

<details>
  <summary>Heavy `cells` function usage example</summary>

  ```javascript
  new Handsontable(example, {
    data: Handsontable.helper.createSpreadsheetData(1000, 500),
    rowHeaders: true,
    colHeaders: true,
    cells(row, col) {
      const cellProperties = {};
      const visualRowIndex = this.instance.toVisualRow(row);
      const visualColIndex = this.instance.toVisualColumn(col);
  
      const data = this.instance.getDataAtCell(visualRowIndex, visualColIndex);
      const dataCol2 = this.instance.getDataAtRowProp(visualRowIndex, 2);
      const dataCol3 = this.instance.getDataAtRowProp(visualRowIndex, 3);
      const dataCol4 = this.instance.getDataAtRowProp(visualRowIndex, 4);
      const date1 = moment();
      const dateFormat1 = moment().format('DD MM YYYY');
      const date2 = moment();
      const dateFormat2 = moment().format('DD MM YYYY');
  
      if (visualRowIndex === 0 && visualColIndex === 0 || date1.isBefore(date2)) {
        cellProperties.readOnly = true;
      }
  
      if (data.indexOf('5') >= 0) {
        cellProperties.className = 'red-bg';
      }
  
      if (row % 5 === 1) {
        cellProperties.className = 'yellow-bg';
      }
  
      if (col % 5 === 1) {
        cellProperties.className = 'blue-bg';
      }
  
      if (row % 12 === 1) {
        cellProperties.readOnly = true;
      }
  
      if (row % 7 === 1 && date1) {
        cellProperties.renderer = function(hot, td, row, col, prop, value) {
          td.textContent = `.x${value}x.`;
        };
      }

      return cellProperties;
    },
  });
  ```
</details>

*All demos were tested using dataset size equals 1000x500 and viewport size set as 1000x800px.

**In conclusion, the increase in performance should be noticeable. Depends on the use case the overall render execution time is improved from 0% to ~50%. The same applies to the `getCellMeta` method. By making direct calls I noticed an improvement of about 50% comparing to the 9.0.2.**

### How has this been tested?
<!--- Please describe in detail how you tested your changes (doesn't apply to translations). -->
I covered the changes with the unit and E2E tests. I tested the changes using Perf Lab.

### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature or improvement (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Additional language file or change to the existing one (translations)

### Related issue(s):
1. Fixes #6303
2.
3.

### Affected project(s):
- [x] `handsontable`
- [x] `@handsontable/angular`
- [x] `@handsontable/react`
- [x] `@handsontable/vue`

### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I have reviewed the guidelines about [Contributing to Handsontable](https://github.com/handsontable/handsontable/blob/master/CONTRIBUTING.md) and I confirm that my code follows the code style of this project.
- [x] I have signed the [Contributor License Agreement](https://docs.google.com/forms/d/e/1FAIpQLScpMq4swMelvw3-onxC8Jl29m0fVp5hpf7d1yQVklqVjGjWGA/viewform?c=0&w=1)
- [ ] My change requires a change to the documentation.