@wojciechczerniak 2021-08-18
Can you attach a graph with performance difference presentation? Not necessarily perf-lab output, I'm looking for something we could use with @kirszenbaum for blog post. May be rough, a graphic designer will fix colors and details.