@budnix 2021-08-19
> Can you attach a graph with performance difference presentation? Not necessarily perf-lab output, I'm looking for something we could use with @kirszenbaum for blog post. May be rough, a graphic designer will fix colors and details.

I thought about that but in the end, I understood that it would be difficult to show it on the chart. Comparing to the suspend rendering feature there was easy to do. There I could put new methods invocation line by line between suspends a could do performance tests and see the difference, it scaled up nicely. Here, I can put something into the `cells` plus add more API calls within but it still depends on the demo. It's very floating. 

Generally, the overall performance of the `getCellMeta` calls increased up to 50%, and thanks to that whole table speed up more or less up to 50%.