@krzysztofspilka 2021-08-19
@budnix Any sort of graphic presentation would definitely help us better communicate the change. Even if that's not possible we need to explain in detail in what conditions we achieved the increased performance, namely:

- What was put in the configuration of the grid that we tested? Were there any custom components included?
- What was the environment used for tests?
- How many repetitions we did to confirm that the change is not a false-positive?
- What measure increased that allows us to claim that there is a gain in performance? First load? Re-render? Scroll efficiency (if so, then how exactly we measured that?)

If not a graph, then maybe we could create two Handsontable "litmus paper" examples next to each other (two code sandbox iframes) to serve as a benchmark inside of the blog post? wdyt?