@budnix 2021-08-20
@wojciechczerniak While collecting the perf data for the chart I found a little bottleneck in our local hook inter-module comm system (high GC activity). Within the one `runLocalHooks` method I found 3 things that cause the GC to run, I fixed that. So, I need your review again.

For basic demo (only headers on) with a dataset 2000x500, now I've got:
Test case 1: `getCellMeta` 25 000 calls (500 rows x 50 columns) each sample
```
9.0.2                                  7.35 ms +-4%  GC time 0.00 ms
10.0.0 (before optim. local hook)      7.63 ms +-4%  GC time 0.12 ms +-235%
10.0.0 (after optim. local hook)       6.87 ms +-2%  GC time 0.00 ms
```
Test case 2: `getCellMeta` 100 000 calls (2000 rows x 50 columns) each sample
```
9.0.2                           28.76 ms +-1%  GC time 0.00 ms
10.0.0 (before local hook opt)  31.12 ms +-1%  GC time 0.37 ms +-77%
10.0.0 (after local hook opt)   26.31 ms +-1%  GC time 0.00 ms
```

I'm still in the phase of collecting more data. Seeing that I'm more than happy with the results.