@krzysztofspilka 2021-08-20
@kirszenbaum should look into this as well. I spotted one or two typos. Nothing important, but it's worth looking into it.